Moving inc.js and add Sound() class in inc.js
